CDH-52469. Fix dependencies to avoid re-packaging Hadoop classes.
It seems maven pulls some dependencies of test jars in compile scope,
and because they weren't declared in the root pom, they get included
in the final assembly; so we ended up with duplicate Hadoop and HDFS
classes.

Instead of declaring these in the root pom, I chose a more surgical
spot and just excluded all Hadoop dependencies from the lineage
module when packaging the assembly. And also added a test to make
sure a few classes are not in the assembly when building the CDH
packaging.