CLOUDERA-BUILD. Hive related modifications in code for Spark2.2.0.cloudera1 to make spark 2.2 compatible with CDH
There are changes in Spark 2.2 upstream that cuase build failures in  Spark2.2.0.cloudera1. The main change is that in upstream Spark 2.2, org.apache.spark.sql.catalyst.catalog.CatalogStorageFormat class uses option[URI] for locationUrithe instead of option[String] in 2.1. Because of these changes, in this commit I made some modifications to downstream Spark 2.2.