CLOUDERA-BUILD. CDH-54858.  Fix blacklist timeout in blacklist preview.
In the version of blacklisting that only exists in CDH 5.9.0 - 5.9.2
(not upstream anywhere), sometimes applying the timeouts results in
calling head on an empty list.  The conditions under which this happens
are really convoluted, but the fix is easy in any case, just make that
call robust to empty lists.