CDH-53384. Add Spark config for choosing locality of the YARN AM.
This is the Spark 2 version of CDH-38538, base on the same code from
MR (MAPREDUCE-6871). The main different here is that because the Spark 2
build needs to work on multiple versions of CDH, this code uses reflection
to call the new YARN API that implements the feature, and falls back to
the old API when it's not present.

(cherry picked from commit f377a6ee14acd68c0130f42e290e078d3f535450)