CLOUDERA_BUILD. Disable Parquet filter push-down for strings.
Spark 2.2 upstream had problems with parquet filter push-down for strings in SPARK-17213. The problem is resolved in upstream by upgrading the parquet version to 1.8.2. We are depending on CDH5.12.0 parquet which doesn't have the required fixes in 1.8.2. As a result we should disable filter push downs for strings. This change also disables the related tests that were enabled after SPARK-17213 was resolved.