CLOUDERA-BUILD. Update gbn resolution URL
In c6 we get the build number from a 'bquery' or an 'alias'. The current
URL represents a 'bquery' for the latest official 6.x build. Using a
well known alias (the branchname) simplifies the branching process and
allows us to point to different bequeries without the need of code
changes.

(cherry picked from commit a04be506972aebeb7dec985d032df96126207907)

Change-Id: Id2d93977f4898d2707170b3997cfd28e7cdbda76