CLOUDERA-BUILD. Update gbn resolution URL
In c6 we get the build number from a 'bquery' or an 'alias'. The current
URL represents a 'bquery' for the latest official 6.x build. Using a
well known alias (the branchname) simplifies the branching process and
allows us to point to different bequeries without the need of code
changes.