CDH-57751. Make sure hive-exec is at end of Spark classpath.
CDH's hive-exec brings way too many shaded libraries that may interfere
with the versions Spark actually uses. Placing it last in the classpath
ensures that Spark sees its own versions of the libraries first.

This is done by packaging the hive-exec jar in a separate directory, and
adding it to the classpath after the main jars directory.