CDH-59241. Fix the "hive.metastore.uris" value that was changed by the backport of SPARK-17531 to 1.6 branches.
The backport of SPARK-17531 to 1.6 branches had set "hive.metastore.uris" in a way that caused an issues with regards to connecting to hive meta store and retrieving hive UDFs. In this change the problematic part of the backport is removed. This fix will be applied to 5.10.x and later branches.

(cherry picked from commit 0d7fa486ff0489ee8f1a799a202a11dfc6e4a727)