CDH-59241. Fix the "hive.metastore.uris" value that was changed by the backport of SPARK-17531 to 1.6 branches.
The backport of SPARK-17531 to 1.6 branches had set "hive.metastore.uris" in a way that caused an issues with regards to connecting to hive meta store and retrieving hive UDFs. In this change the problematic part of the backport is removed. This fix will be applied to 5.10.x and later branches.

(cherry picked from commit 7c1295aa76d587d577b27d2fb0ccea3ad590bcc4)