CDH-62252. Convert Impala-written timestamps from UTC to local TZ (SPARK-12297)
Parquet Int96 timestamp data written by Impala vs data written by Hive
& Spark is stored slightly differently -- they use a different offset
for the timezone. This adds an option
"spark.sql.parquet.int96TimestampConversion" (false by default) to
adjust timestamps if and only if the writer is Impala (or more
precisely, if the parquet file's "createdBy" metadata does not start
with "parquet-mr").  This matches the existing behavior in Hive from
HIVE-9482.

The implementation is different from the upstream Spark 2.x version,
as the Parquet reader changed significantly when vectorized readers
were added. Here, we rely on Parquet's 'read support' classes to
inject the timestamp conversion code, and plumb configuration data
through to the CatalystRowConverter that implements the conversion via
a Configuration object. There is no change to the read path when
conversion is disabled; there are no new per-value branches in either
case.

The test is broadly the same as upstream, except that vectorization is
not supported, so we remove that test axis, and an upstream assumption
that is tested about column stats is not valid in this version.