[SPARK-23436][SQL][BACKPORT-2.3] Infer partition as Date only if it can be casted to Date
This PR is to backport https://github.com/apache/spark/pull/20621 to branch 2.3

---

Before the patch, Spark could infer as Date a partition value which cannot be casted to Date (this can happen when there are extra characters after a valid date, like `2018-02-15AAA`).

When this happens and the input format has metadata which define the schema of the table, then `null` is returned as a value for the partition column, because the `cast` operator used in (`PartitioningAwareFileIndex.inferPartitioning`) is unable to convert the value.

The PR checks in the partition inference that values can be casted to Date and Timestamp, in order to infer that datatype to them.

added UT

Author: Marco Gaido <marcogaido91@gmail.com>

Closes #20764 from gatorsmile/backport23436.

(cherry picked from commit 8ff8e16e2da4aa5457e4b1e5d575bd1a3a1f0358)